+++ packages/fighting-design/popup/src/interface.d.ts
@@ -2,6 +2,6 @@ export type { PopupPropsType } from './popup'
 
 export type PopupDirectionType = 'left' | 'right' | 'top' | 'bottom' | 'center'
 
-export interface PopupHandleChangeInterface {
+export interface PopupCallbackInterface {
   (evt: MouseEvent): void
 }