+++ packages/fighting-design/popup/src/popup.vue
@@ -4,7 +4,7 @@
   import { sizeChange } from '../../_utils'
   import type { ComputedRef, CSSProperties } from 'vue'
   import type { OrdinaryFunctionInterface as a } from '../../_interface'
-  import type { PopupHandleChangeInterface as b } from './interface'
+  import type { PopupCallbackInterface as b } from './interface'
 
   const prop = defineProps(Props)
   const emit = defineEmits(Emits)