+++ CHANGELOG.md
@@ -2,10 +2,12 @@
 
 中文 | [英文](https://github.com/FightingDesign/fighting-design/blob/master/CHANGELOG.en-US.md)
 
+## 0.7.1-alpha.1 (2022-10-25)
+
 - 新增 `f-swap` 组件
 - 修复 `f-image-preview` 组件预览旋转方向错误的问题
 - 优化 `f-image-preview` 组件内部性能
-- `r-popup` 组件 `open` `open-end` `close` `close-end` 改为 `props` 参数
+- `f-popup` 组件 `open` `open-end` `close` `close-end` 改为 `props` 参数
 
 ## 0.7.0-beta.2 (2022-10-22)